unemployment_world
Background:
In this case study, trainees will explore a dataset that captures unemployment rates across different countries from 1991 to 2021. The dataset provides an excellent opportunity to analyze global economic trends, understand regional disparities, and assess the impact of major global events on employment.

Problem Statement
What is the problem?
Analyze the global unemployment trends over three decades to identify patterns, anomalies, and the impact of major economic events (like the 2008 financial crisis and the 2020 COVID-19 pandemic).

Why is it important to solve it?
Understanding unemployment trends is crucial for policymakers, businesses, and international organizations to devise strategies for economic growth, workforce development, and crisis management.
Data Dictionary
Country Name: Name of the country.
Country Code: Three-letter country code.
Years (1991 to 2021): Unemployment rates for each year, are represented as a percentage of the total labor force.