mobile_apps_playstore_dataanalysis
Why this analysis - This Data Analysis is for getting insights about the mobile apps w.r.t its future business approaches and development of mobile apps. This study has been summarised from the historical data available from the Google Play store from June 2012 till February 2019. It also helps to know the trend involved in the categories of mobile app availability, usability, and users' expectations.

What data is available - The above dataset contains information such as names of the mob apps from the Google Play store, categories, Android versions, Rating, Number of reviews, Number of installs, size of the app, when was last updated, and the current version of the mob apps.

Recommended Analysis
Q1: - Which is the most used Android version?
Q2: - Which all categories of mob apps are existing from this dataset?
Q3: - Which is the most installed category of mob app from this dataset?
Q4: - How many total categories are there?
Q5: - Which apps have got 5-star ratings
Q6: - Which are the top 10 mob apps based on ratings?