nifty50_stock_prices
Historical stock market data for current Nifty 50 companies, from 2000-2021. Each record represents a single day of trading and includes the ticker name, volume, and high, low, open, and close prices.
Recommended Analysis
1.Ranking overall, by Industry and by single title (top 5), using mean Volume (and Trades)
Group by Industry - Calculate the Mean of Volume and Trades - Top 5 Values 
Group by Ticker - Calculate the Mean of Volume and Trades - Top 5 Values 
2.Ranking of Must trading stock Since Inception (using Trades)
Group by Ticker - Calculate the Mean and Sum of Trades - Top 5 Values 
3.Top 5 best stocks using delta price % Since Inception (https://sciencing.com/calculate-delta-percentage-8475192.html)	
GroupBy on the basis of the company name
4.Calculate the percentage change using the equation [X(final) - X(initial)] / X(initial) * 100.
Final - the last occurrence of close
Initial - the first occurrence of close
5.Top 5 worst stocks using delta price % Since Inception
6.Standard deviation is a statistical term that measures the amount of variability or dispersion around an average. Standard deviation is also a measure of volatility: using Numpy is easy to calculate it for each stock or sector into Index.
With SD, we could calculate the coefficient of variation (CV=standard deviation/mean). As a rule of thumb, a CV >= 1 indicates a relatively high variation, while a CV < 1 can be considered low. This means that distributions with a coefficient of variation higher than 1 are considered to be high variance whereas those with a CV lower than 1 are considered to be low variance.